Svc REST: Move isValidAddress to storage.go
